+++ Changelog.md
@@ -14,6 +14,10 @@ Der entsprechende Abschnitt des Changelogs wird auch jeweils in die [Releasenote
 
 ## Unreleased
 
+### Änderungen
+- Export in Clipboard anpassen, Einstellungen für CSV Export, Übersetzungen ab i18n 12.27.0
+
+
 ## [12.14.1] - 15.01.2025
 
 ### Änderung